requirements.txt
python-binance==1.0.19
rich==13.7.1